Added a basic server, client and debug file
The server talks to one SBHS only for now.

The client is a command line client where JSON values
have to be entered on a command line.

The debug file (sbhs_dummy.py) enabled us to test the
server when the actual SBHS is not working or is
not connected.